Refactor: folder name & delete redundant ContentSection component
